pyproject.toml
[tool.poetry]
name = "ladsetal"
version = "0.1.0"
description = "the blog"
authors = ["Ben Simms <ben@bensimms.moe>"]

[tool.poetry.dependencies]
python = "^3.8"
itsdangerous = "^1.1"
cryptography = "^3.3"
gino = "^1.0"
python-dotenv = "^0.15"
sqlalchemy_utils = "^0.36"
asyncpg = "^0.21"
alembic = "^1.4"
"sqlalchemy-searchable" = { tag = "1.1.2", git = "https://github.com/nitros12/sqlalchemy-searchable" }
starlette = "^0.14"
jinja2 = "^2.11"
python-multipart = "^0.0"
pygments = "^2.6"
aiofiles = "^0.6"
slug = "^2.0"
WTForms = "^2.3"
yarl = "^1.6"
gunicorn = "^20.0"
PyCrypto = "^2.6"
python-dateutil = "^2.8"
parsedatetime = "^2.6"
uvicorn = "^0.13"
psycopg2-binary = "^2.8.6"
mistune = "2.0.0a6"
authlib = {git = "https://github.com/lepture/authlib.git"}
httpx = "^0.16.1"
orjson = "^3.4.6"

[tool.poetry.dev-dependencies]
psycopg2-binary = "^2.8"

[tool.poetry.scripts]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.core.masonry.api"
